move search helpers out of views
